# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "For the Cult API",
    "description": "The most AI-agent-friendly eCommerce API for quality lifestyle products. Designed for seamless integration with AI assistants, chatbots, and autonomous agents. Features semantic search, structured errors with suggestions, and multi-chain crypto payments.",
    "version": "1.0",
    "contact": {
      "name": "For the Cult",
      "url": "https://forthecult.store"
    },
    "x-ai-friendly": true,
    "x-quick-start": "Call GET /api/agent/capabilities first to understand the API"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base (relative)"
    }
  ],
...
  "paths": {
...
    "/orders/{orderId}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get order details",
        "description": "Full order: items, shipping, payment summary.",
        "operationId": "getOrderById",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full order",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orderId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "paidAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    },
                    "email": {
                      "type": "string"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "shipping": {
                      "type": "object"
                    },
                    "totals": {
                      "type": "object"
                    },
                    "payment": {
                      "type": "object"
                    },
                    "_actions": {
                      "type": "object",
                      "description": "Action hints for AI agents",
                      "properties": {
                        "next": {
                          "type": "string"
                        },
                        "cancel": {
                          "type": "string"
                        },
                        "help": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    }
  },
...
}